Rework model classes
So far, the model classes were also view models. I want to get rid of that and introduce specific view model classes to wrap those business models in the UI layer.
This should make it easier to transition to a .NET Standard 2.0 Logic project, as I don't rely on Bitmap, Color and the like.